Corrected package name to be lowercase.
