--- src/main/java/com/QualityQuizzes/Quiz/QuizApplication.java
@@ -1,13 +1,0 @@
-package com.QualityQuizzes.Quiz;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class QuizApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(QuizApplication.class, args);
-	}
-
-}
