--- src/main/java/com/QualityQuizzes/Quiz/controller/QuizController.java
@@ -1,12 +1,0 @@
-package com.QualityQuizzes.Quiz.controller;
-
-import org.springframework.web.bind.annotation.CrossOrigin;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@CrossOrigin(origins = "http://localhost:8080")
-@RestController
-@RequestMapping("/api")
-public class QuizController {
-    // Methods ////////////////////////////////////////////////////////////////////////////////////////////////////////
-}
