--- src/main/java/com/QualityQuizzes/Quiz/mapper/QuizQuestionMapper.java
@@ -1,23 +1,0 @@
-package com.QualityQuizzes.Quiz.mapper;
-
-import com.QualityQuizzes.Quiz.model.QuizQuestion;
-import org.springframework.jdbc.core.RowMapper;
-
-import java.sql.ResultSet;
-import java.sql.SQLException;
-
-
-public class QuizQuestionMapper implements RowMapper<QuizQuestion> {
-    
-    @Override
-    public QuizQuestion mapRow(ResultSet rs, int rowNum) throws SQLException {
-        QuizQuestion quizQuestion = new QuizQuestion();
-        
-        quizQuestion.setQuestion        (rs.getString("question"));
-        quizQuestion.setIncorrectAnswer (rs.getString("incorrectAnswer"));
-        quizQuestion.setCorrectAnswer   (rs.getString("correctAnswer"));
-        quizQuestion.setId              (rs.getLong("id"));
-        
-        return quizQuestion;
-    }
-}
