+++ src/main/java/com/qualityquizzes/quiz/QuizApplication.java
@@ -1,0 +1,13 @@
+package com.qualityquizzes.quiz;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class QuizApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(QuizApplication.class, args);
+	}
+
+}
