+++ src/main/java/com/qualityquizzes/quiz/controller/QuizController.java
@@ -1,0 +1,12 @@
+package com.qualityquizzes.quiz.controller;
+
+import org.springframework.web.bind.annotation.CrossOrigin;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@CrossOrigin(origins = "http://localhost:8080")
+@RestController
+@RequestMapping("/api")
+public class QuizController {
+    // Methods ////////////////////////////////////////////////////////////////////////////////////////////////////////
+}
